refactor(infrastructure): split Drizzle schema per bounded context (closes #102)
Drizzle-kit 0.31.10 resolves cross-file `.js` ESM imports between sibling
schema modules — the comment in the old monolithic `schema/index.ts`
claiming otherwise was stale. Empirically verified that splitting the 1175
LOC monolith into 14 per-context files (one per bounded context plus a
forward-looking engagement.ts for the unwired Brevo tables) produces
zero migration diff against the live schema.

Each schema file now exports its tables as a typed `<context>SchemaTables`
const so per-context modules can populate `ContextRegistrations.schemaTables`
without crossing the application/infrastructure layer boundary.

The committed `0015_snapshot_reconciliation.sql` is a one-time idempotent
migration: hand-written 0013/0014 never wrote `meta/0013_snapshot.json` /
`meta/0014_snapshot.json`, so the model drifted. The new migration is
guarded with `IF NOT EXISTS` / `EXCEPTION duplicate_object` so it's a
no-op on databases that already ran 0013 and a verbatim apply on fresh
databases. Subsequent `drizzle-kit generate` runs are clean.
